+++ docs/versionhistory.rst
@@ -9,6 +9,7 @@ This library adheres to `Semantic Versioning <http://semver.org/>`_.
 - **BACKWARD INCOMPATIBLE** Renamed the ``asphalt.templating.util`` module to
   ``asphalt.templating.utils``
 - Renderer resources are now added to the context using their actual types as well
+- The Mako renderer now skips filesystem checks by default in production mode
 
 **1.0.0** (2015-05-21)